New design with user commands
